requirements.txt
geopandas==0.12.1
matplotlib==3.6.2
numpy==1.21.5
pandas==1.4.4
PyYAML==6.0
seaborn==0.12.1
